requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
dash==2.18.2
